Refactoring: productDetaio페이지 디자인패턴 적용
